# backend/internal/repository/sqlc/models.go
// Code generated by sqlc. DO NOT EDIT.
// versions:
//   sqlc v1.30.0

package sqlc

import (
	"github.com/google/uuid"
	"github.com/jackc/pgx/v5/pgtype"
)

type Dependency struct {
	ID         uuid.UUID `json:"id"`
	ProjectID  uuid.UUID `json:"projectId"`
	FromTaskID uuid.UUID `json:"fromTaskId"`
	ToTaskID   uuid.UUID `json:"toTaskId"`
}

type Invite struct {
	Token     string             `json:"token"`
	ProjectID uuid.UUID          `json:"projectId"`
	CreatedBy uuid.UUID          `json:"createdBy"`
	ExpiresAt pgtype.Timestamptz `json:"expiresAt"`
	CreatedAt pgtype.Timestamptz `json:"createdAt"`
}

type Project struct {
	ID        uuid.UUID          `json:"id"`
	Title     string             `json:"title"`
	OwnerID   uuid.UUID          `json:"ownerId"`
	CreatedAt pgtype.Timestamptz `json:"createdAt"`
}

type ProjectMember struct {
	ProjectID uuid.UUID `json:"projectId"`
	UserID    uuid.UUID `json:"userId"`
	Role      string    `json:"role"`
}

type Task struct {
	ID         uuid.UUID          `json:"id"`
	ProjectID  uuid.UUID          `json:"projectId"`
	Title      string             `json:"title"`
	Status     string             `json:"status"`
	AssigneeID *uuid.UUID         `json:"assigneeId"`
	X          float64            `json:"x"`
	Y          float64            `json:"y"`
	UpdatedAt  pgtype.Timestamptz `json:"updatedAt"`
	CreatedAt  pgtype.Timestamptz `json:"createdAt"`
}

type User struct {
	ID           uuid.UUID          `json:"id"`
	Email        string             `json:"email"`
	PasswordHash string             `json:"passwordHash"`
	DisplayName  string             `json:"displayName"`
	CreatedAt    pgtype.Timestamptz `json:"createdAt"`
}